+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 0.4.0 / 2014-12-26
+
+* [FEATURE] specify default decorators in a class with .decorators
+* [FEATURE] threadsafe shared state (w/ documentation)
+
 # 0.3.0 / 2014-2-3
 
 * [FEATURE] directly callable decorators